OPENEUROPA-1551: Separate logger subscriber from logger.
